// apps/dashboard/src/generated/graphql.tsx
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  Date: { input: any; output: any; }
  JSON: { input: any; output: any; }
};
// ...
export type Constellation = {
  __typename?: 'Constellation';
  Satellites: Maybe<Array<Maybe<Satellite>>>;
  description: Scalars['String']['output'];
  id: Scalars['ID']['output'];
  name: Scalars['String']['output'];
};
// ...
export type Contact = {
  __typename?: 'Contact';
  Employee: Maybe<Employee>;
  GroundStation: Maybe<GroundStation>;
  Payload: Maybe<Payload>;
  Satellite: Maybe<Satellite>;
  configuration: Scalars['JSON']['output'];
  date: Scalars['Date']['output'];
  employee_id: Scalars['ID']['output'];
  executionScript: Scalars['String']['output'];
  groundStation_id: Scalars['ID']['output'];
  id: Scalars['ID']['output'];
  payload_id: Maybe<Scalars['ID']['output']>;
  satellite_id: Scalars['ID']['output'];
  type: Scalars['String']['output'];
};
// ...
export type Launch = {
  __typename?: 'Launch';
  Satellites: Maybe<Array<Maybe<Satellite>>>;
  date: Scalars['Date']['output'];
  id: Scalars['ID']['output'];
  outcome: Maybe<Scalars['String']['output']>;
  provider: Scalars['String']['output'];
  rocket: Scalars['String']['output'];
  status: Scalars['String']['output'];
};
// ...
export type Payload = {
  __typename?: 'Payload';
  Contacts: Maybe<Array<Maybe<Contact>>>;
  Customer: Maybe<Customer>;
  Satellite: Maybe<Satellite>;
  category: Scalars['String']['output'];
  configuration: Scalars['JSON']['output'];
  customer_id: Scalars['ID']['output'];
  description: Scalars['String']['output'];
  id: Scalars['ID']['output'];
  name: Scalars['String']['output'];
  satellite_id: Scalars['ID']['output'];
  status: Scalars['String']['output'];
};
// ...
export type Report = {
  __typename?: 'Report';
  Comments: Maybe<Array<Maybe<Comment>>>;
  Employee: Maybe<Employee>;
  GroundStation: Maybe<GroundStation>;
  Satellite: Maybe<Satellite>;
  content: Scalars['String']['output'];
  date: Scalars['Date']['output'];
  employee_id: Scalars['ID']['output'];
  groundStation_id: Maybe<Scalars['ID']['output']>;
  id: Scalars['ID']['output'];
  satellite_id: Maybe<Scalars['ID']['output']>;
  title: Scalars['String']['output'];
  type: Scalars['String']['output'];
};
// ...
export type Satellite = {
  __typename?: 'Satellite';
  Constellation: Maybe<Constellation>;
  Contacts: Maybe<Array<Maybe<Contact>>>;
  Launch: Maybe<Launch>;
  Payloads: Maybe<Array<Maybe<Payload>>>;
  Reports: Maybe<Array<Maybe<Report>>>;
  altitude: Scalars['Float']['output'];
  busType: Scalars['String']['output'];
  constellation_id: Maybe<Scalars['ID']['output']>;
  coordinates: Array<Maybe<Scalars['Float']['output']>>;
  description: Scalars['String']['output'];
  id: Scalars['ID']['output'];
  image: Scalars['String']['output'];
  launch_id: Scalars['ID']['output'];
  manufacturer: Scalars['String']['output'];
  name: Scalars['String']['output'];
  specs: Scalars['JSON']['output'];
  status: Scalars['String']['output'];
  tle: Scalars['JSON']['output'];
};